Floundering around. All previous cal eg calmon42x provided gain labels from xml and used xml cal values to modify ESLO, EOFF
The plan for ACQ426 was different. The ADI chip driver already provides
the labels, and it's supposed to make a live cal adjust.
Live cal adjust is probably tricky to achieve.

Could we go back to the old strategy of "let the user do it".
That's the goal of this switch
SUBCALMON42x_FOR_CALMON426=y

Unfortunately it doesn't work. caldef "sw" keys were always single digit
integers, and the embedded xml processing assumes this. However for
ACQ426, sw="LABEL", so that fails. Maybe sw could go back to being an
ordinal. However, that didn't seem to work.

Out of time, park it @@todo
 On branch epics7
	modified:   scripts/load.records